# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bindizr HTTP API
  description: This is the API documentation for Bindizr
  contact:
    email: kevin136583@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0-beta.7
paths:
...
  /records:
    get:
      tags:
      - Record
      summary: List all DNS records
      operationId: list_records
      parameters:
      - name: zone_name
        in: query
        description: The name of the DNS zone to filter records by.
        required: false
        schema:
          type: string
      - name: name
        in: query
        description: Filter by record name.
        required: false
        schema:
          type: string
      - name: record_type
        in: query
        description: Filter by record type.
        required: false
        schema:
          type: string
      - name: value
        in: query
        description: Partially filter by record value.
        required: false
        schema:
          type: string
      - name: ttl
        in: query
        description: Filter by TTL.
        required: false
        schema:
          type: integer
          format: int32
      - name: min_ttl
        in: query
        description: Filter by minimum TTL.
        required: false
        schema:
          type: integer
          format: int32
      - name: max_ttl
        in: query
        description: Filter by maximum TTL.
        required: false
        schema:
          type: integer
          format: int32
      - name: priority
        in: query
        description: Filter by priority.
        required: false
        schema:
          type: integer
          format: int32
      - name: min_priority
        in: query
        description: Filter by minimum priority.
        required: false
        schema:
          type: integer
          format: int32
      - name: max_priority
        in: query
        description: Filter by maximum priority.
        required: false
        schema:
          type: integer
          format: int32
      - name: search
        in: query
        description: Partially search records.
        required: false
        schema:
          type: string
      - name: sort
        in: query
        description: Sort by name (the default), record_type, ttl, priority, or created_at.
        required: false
        schema:
          type: string
      - name: order
        in: query
        description: asc (the default) or desc.
        required: false
        schema:
          type: string
      - name: signed
        in: query
        description: Append the zone's derived DNSSEC records (RRSIG, DNSKEY, NSEC/NSEC3/NSEC3PARAM, CDS, CDNSKEY) after the user records, in the same pagination. Derived rows carry no id, and record_type also accepts a derived type. A search narrows them by name only — their type is stored as a number and their rdata as wire bytes — a priority filter leaves them out, since none carries one, and a value filter is refused outright rather than answered without them.
        required: false
        schema:
          type: boolean
      - name: limit
        in: query
        description: Records per page; defaults to 50.
        required: false
        schema:
          type: integer
          format: int32
          maximum: 1000
          minimum: 1
      - name: offset
        in: query
        description: Number of records to skip.
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      responses:
        '200':
          description: A list of DNS records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_GetRecordResponse'
        '400':
          description: Bad request, invalid pagination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - Record
      summary: Create a new DNS record
      operationId: create_record
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecordRequest'
        required: true
      responses:
        '201':
          description: DNS record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordResponse'
        '400':
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The token's grants do not allow this record write
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Zone not found, or not visible to the token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Record conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type, expected JSON request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
      - Record
      summary: Delete records by name
      description: 'Removes every record matching the filter in one transaction, so the zone advances by a single serial and sends one NOTIFY. Narrowing follows RFC 2136, Section 2.5.2: a name alone takes every type at it, adding record_type narrows to that type, adding value takes one record. Matching nothing is not an error — the zone already reads the way the request asked for, so nothing moves.'
      operationId: delete_records_matching
      parameters:
      - name: zone_name
        in: query
        required: true
        schema:
          type: string
      - name: name
        in: query
        description: Owner name relative to the zone, or `@` for the apex.
        required: true
        schema:
          type: string
      - name: record_type
        in: query
        description: Narrows to one type; omitted, every type at the name goes.
        required: false
        schema:
          type: string
      - name: value
        in: query
        description: |-
          Narrows to one record; compared canonically, so a value spelled
          another way still matches. Requires `record_type`.
        required: false
        schema:
          type: string
      - name: priority
        in: query
        description: |-
          MX and SRV keep their preference in its own column, so it narrows
          there rather than being part of `value`.
        required: false
        schema:
          type: integer
          format: int32
      - name: dry_run
        in: query
        description: When true, report what would go without removing anything.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Records deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteRecordsResponse'
        '400':
          description: Invalid filter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Zone not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CreateRecordRequest:
      type: object
      description: Request body for creating a record in a named zone.
      required:
      - name
      - record_type
      - value
      - zone_name
      properties:
        name:
          type: string
          example: sub
        priority:
          type:
          - integer
          - 'null'
          format: int32
          description: MX and SRV priority, set here rather than inline in the value; other record types reject it.
          example: 10
        record_type:
          type: string
          example: A
        ttl:
          type:
          - integer
          - 'null'
          format: int32
          description: Optional; an omitted TTL is fixed to the zone's TTL at write time. Records sharing a name and type share one TTL.
          example: 3600
        value:
          $ref: '#/components/schemas/RecordValueRequest'
        zone_name:
          type: string
          example: example.com
...
    DeleteRecordsResponse:
      type: object
      description: |-
        What a conditional delete removed, or would have. Matching nothing is not
        an error: the zone already reads the way the request asked for, so the
        serial does not move and no NOTIFY goes out.
      required:
      - applied
      - dry_run
      - deleted
      - records
      - diff
      properties:
        applied:
          type: boolean
          example: true
        deleted:
          type: integer
          example: 3
          minimum: 0
        diff:
          $ref: '#/components/schemas/RecordDiff'
          description: The removal as a record diff, for previewing the change.
        dry_run:
          type: boolean
          example: false
        records:
          type: array
          items:
            $ref: '#/components/schemas/GetRecordResponse'
...
    ErrorResponse:
      type: object
      description: 'Generic error response: a plain description plus a machine-readable code.'
      required:
      - error
      - code
      properties:
        code:
          type: string
          example: ZONE_NOT_FOUND
        error:
          type: string
          example: Zone with name 'example.com' not found
...
    GetRecordResponse:
      type: object
      description: API representation of a record.
      required:
      - name
      - record_type
      - value
      - ttl
      - zone_id
      - zone_name
      properties:
        id:
          type:
          - integer
          - 'null'
          format: int32
          description: |-
            Absent on the derived DNSSEC rows of a signed listing, which are not
            addressable records.
          example: 1
        name:
          type: string
          example: sub
        priority:
          type:
          - integer
          - 'null'
          format: int32
          example: 10
        record_type:
          type: string
          example: A
        ttl:
          type: integer
          format: int32
          example: 3600
        value:
          $ref: '#/components/schemas/RecordValueRequest'
        zone_id:
          type: integer
          format: int32
          example: 1
        zone_name:
          type: string
          example: example.com
...
    PaginatedResponse_GetRecordResponse:
      type: object
      description: A page of items together with its pagination metadata.
      required:
      - items
      - pagination
      properties:
        items:
          type: array
          items:
            type: object
            description: API representation of a record.
            required:
            - name
            - record_type
            - value
            - ttl
            - zone_id
            - zone_name
            properties:
              id:
                type:
                - integer
                - 'null'
                format: int32
                description: |-
                  Absent on the derived DNSSEC rows of a signed listing, which are not
                  addressable records.
                example: 1
              name:
                type: string
                example: sub
              priority:
                type:
                - integer
                - 'null'
                format: int32
                example: 10
              record_type:
                type: string
                example: A
              ttl:
                type: integer
                format: int32
                example: 3600
              value:
                $ref: '#/components/schemas/RecordValueRequest'
              zone_id:
                type: integer
                format: int32
                example: 1
              zone_name:
                type: string
                example: example.com
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      description: Pagination window and total count for a list response.
      required:
      - limit
      - offset
      - total
      properties:
        limit:
          type: integer
          format: int32
          example: 50
          minimum: 0
        offset:
          type: integer
          format: int64
          example: 0
          minimum: 0
        total:
          type: integer
          format: int64
          example: 125
          minimum: 0
    RecordDiff:
      type: object
      description: |-
        Record differences grouped by name and type. Version comparisons always
        populate this; mutation responses populate it only for dry-run previews.
      required:
      - entries
      - summary
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/RecordDiffEntry'
        summary:
          $ref: '#/components/schemas/RecordDiffSummary'
    RecordDiffEntry:
      type: object
      description: |-
        The records of one name and type that differ, with those present on
        each side. `from` is empty for `added`, `to` for `removed`.
      required:
      - change
      - name
      - record_type
      - from
      - to
      properties:
        change:
          type: string
          description: '`added`, `removed`, or `changed`.'
          example: changed
        from:
          type: array
          items:
            $ref: '#/components/schemas/RecordDiffValue'
        name:
          type: string
          example: www.example.com.
        record_type:
          type: string
          example: A
        to:
          type: array
          items:
            $ref: '#/components/schemas/RecordDiffValue'
    RecordDiffSummary:
      type: object
      description: How many name-and-type groups of records were added, removed, and changed.
      required:
      - added
      - removed
      - changed
      properties:
        added:
          type: integer
          example: 1
          minimum: 0
        changed:
          type: integer
          example: 1
          minimum: 0
        removed:
          type: integer
          example: 1
          minimum: 0
    RecordDiffValue:
      type: object
      description: |-
        One record on one side of a diff. Rendering (zone-file rdata, priority
        placement) is left to the client; the value is in display form.
      required:
      - value
      - ttl
      properties:
        priority:
          type:
          - integer
          - 'null'
          format: int32
          example: 10
        ttl:
          type: integer
          format: int32
          example: 300
        value:
          $ref: '#/components/schemas/RecordValueRequest'
...
    RecordResponse:
      type: object
      description: A single record wrapped in a response envelope.
      required:
      - record
      properties:
        record:
          $ref: '#/components/schemas/GetRecordResponse'
    RecordValueRequest:
      oneOf:
      - type: string
        example: 192.168.1.100
      - type: array
        items:
          type: string
        example:
        - hello
        - world
      description: 'A record value as sent by the client: a single string or TXT segments.'
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
security:
- bearer_auth: []
...